Modify Solder XAML Update

+++ CadreApp/Solder/CacData.cs
@@ -5,6 +5,6 @@ public class CACData
     public string FirstName { get; set; }
     public string LastName { get; set; }
     public string Rank { get; set; }
-    public int? DODID { get; set; }
+    public long? DODID { get; set; }
     
 }
